refactor: Consolidate byte order helpers and address PR feedback
- Create Common/headers/byteorder_helpers.h with consolidated host/disk
  byte order conversion functions
- Update menupack.c, langhash.c, and memory.c to use shared header
- Remove ~185 lines of duplicated inline helper definitions
- Upgrade truncation warnings from log_debug to log_warn for visibility
- Update test skip reasons to reference correct stub file

This addresses the minor feedback items from the PR #383 code review.
